Avoid loading the wrong mime.types file from the classpath
When there is another JAR on the classpath with /META-INF/mime.types in it, then using .getResourceAsStream will get the first one it finds on the classpath, which might not be the one you want.

"where more than one module contains a resource with the given name, then the ordering that modules are searched is not specified and may be very unpredictable" https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/lang/ClassLoader.html#getResource(java.lang.String)

Choices were either to use .getResources("META-INF/mime.types") and then try and find the path associated with the HTTP4K JAR, or put the mime.types file in a more unique location. Decided to go for the more unique location as it offered more control than hoping to find the right JAR in an unbounded classpath.

Co-authored-by: Diego Gliosca <diego.gliosca@gmail.com>