Fix invalid OpenAPI2 when root and base path match
When generating OpenAPI2 descriptions where the root and base path
math, an invalid description is generated.

For example:

  val router = "/" bind contract {
      routes += "/" bindContract Method.GET to { Response(Status.OK) }
      descriptionPath = "/docs"
  }

Would generate swagger as follows:

  {
    "swagger": "2.0",
    "info": {
      "title": "title",
      "version": "1.2",
      "description": "module description"
    },
    "basePath": "/",
    "paths": {
      "": {
        "get": {
          "operationId": "get"
        }
      }
    }
  }

Which is invalid as empty strings for path URIs is not allowed. This
can be tested by running the output of the API through a swagger
validator as follows:

  $ npm install swagger-cli -g
  $ swagger-cli validate example-http4k-swagger.json