Fix: host should be typed and OpenAPI2 only
As per the comments on the PR here:

https://github.com/http4k/http4k/pull/268#issuecomment-514087853

The host is now passed into the `OpenAPI2` directly via a strongly
typed `baseUri` parameter. The host information is extracted from this
object.

The new parameter is optional to retain backwards-compatibility.

As part of this change I also extracted schema information from the
`baseUri` parameter.

In future, the base path could also be extracted from this parameter.
I have not done this to keep the commit small, and in case the change
breaks backwards-compatibility.