Optionally include description path in description
Google Cloud Endpoints (GCE) will only make an endpoint publicly
accessible if it has a path directive in the OpenAPI description. But
by default http4k-contract does not include the contract path in the
OpenAPI description.

This means hosted Swagger documentation is inaccessible unless we
manually re-format the results of the contract route.

This change adds an optional directive to the contract DSL that adds
the description path to the OpenAPI description, allowing it to be
accessed when deployed in GCE.