Introduce new AppLoaderWithContexts interface
This interface is intended to replace the existing `AppLoader`
interface, adding a new parameter to contain the `RequestContexts`
where lambda-specific additional context can be supplied.

This interface was added in a backwards-compatible way, by keeping the
existing `AppLoader` interface untouched, and adding support for both.

In the future, it may be relevant to consider deprecating the old one.

Issue #287