Use APIGatewayProxy[Request/Response]Event from aws-lambda-java-events
The format of the input message for a lambda is documented in
<https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format>
and several of these fields were missing from the previous
`ApiGatewayProxyRequest`/`ApiGatewayProxyResponse`.

This means that in particular, fields such as the `requestContext`,
and `stageVariables`, as well as multi-value headers and parameters
were unavailable.

Rather than manually adding each of the missing fields (and its
nested fields) to the proxy request/response classes, instead I
went ahead and added a dependency on the small (around 100kb as
of this writing --
<https://mvnrepository.com/artifact/com.amazonaws/aws-lambda-java-events/2.2.7>)
`aws-lambda-java-events` library, which already includes POJOs for
all of the fields and their nested objects.