Graceful shutdown on server stop
Make stop() gracefully shutdown a server without terminating inflight
requests. Decided to take 10 seconds as the default graceful shutdown
timeout for now, but should probably be configurable.

The implementation of the GracefulShutdown-Filter is based on Undertow's
GracefulShutdownHandler.

When a shutdown is initiated the filter will allow the existing requests
to be handled but immediately starts responding with 503 to new
requests. Many HTTP clients will retry on 503 and hopefully arrive at
another server that is not currently shutting down.