update "nestable routes" docs
Variables in paths are covered in the docs but I didn't find anything about restricting them via Regex. I wanted to clarify and add a few examples for people who aren't familiar with Regex's or who might be coming from a framework (Flask, Bottle) which has pre-defined types. Please feel free to edit/modify any of the docs or examples to better match with http4k's existing documentation style. Or, if this is already covered, replace the whole thing with a link.